refactor: fix audit rotation and remove unnecessary conversions
Audit package improvements:
- rotatingWriter.rotate(): return close error instead of ignoring it
- rotatingWriter.cleanup(): split into focused helpers to reduce cyclomatic complexity
- cleanupByAge() and cleanupByCount(): handle stat errors explicitly
- findOldest(): helper for backup limit cleanup
- Use modern octal literals (0o750, 0o600, 0o644)
- Use http.NoBody in test requests

Remove unnecessary conversions found by unconvert:
- auth/dkim.go: ed25519.PrivateKey(privateKey) -> privateKey
- auth/dmarc_report.go: string(adkim) -> adkim, string(aspf) -> aspf
- auth/ldap.go: timeDuration(c.config.Timeout) -> c.config.Timeout
- api/server_accounts.go: string(hashedPassword) -> hashedPassword
- sieve/interpreter.go: int64(t.Size) -> t.Size
- storage/database.go: uint32(btoi(...)) -> btoi(...) where btoi already returns uint32
- cmd/umailclient/main.go: int64(sent) -> sent

All tests pass.

+++ cmd/umailclient/main.go
@@ -269,7 +269,7 @@ func runBench() {
 	fmt.Printf("Time:     %v\n", elapsed)
 	fmt.Printf("Rate:     %.2f msg/sec\n", float64(sent)/elapsed.Seconds())
 	if sent > 0 {
-		fmt.Printf("Avg:      %v\n", time.Duration(totalLat.Nanoseconds()/int64(sent)))
+		fmt.Printf("Avg:      %v\n", time.Duration(totalLat.Nanoseconds()/sent))
 		fmt.Printf("Min:      %v\n", minLat)
 		fmt.Printf("Max:      %v\n", maxLat)
 		if len(latencies) > 0 {
